SRB2: long is not always 32-bit
git-svn-id: svn://code.srb2.org/SRB2/trunk@4500 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/d_player.h
@@ -338,8 +338,8 @@ typedef struct player_s
 	////////////////////
 	// Tag Mode Stuff //
 	////////////////////
-	long tagzone; // Tag Zone timer
-	long taglag; // Don't go back in the tag zone too early
+	INT32 tagzone; // Tag Zone timer
+	INT32 taglag; // Don't go back in the tag zone too early
 
 	////////////////////
 	// CTF Mode Stuff //
